+++ prismedia/upload.py
@@ -11,9 +11,8 @@ Usage:
   prismedia --version
 
 Options:
-  -f, --file=STRING Path to the video file to upload in mp4
+  -f, --file=STRING Path to the video file to upload in mp4. This is the only mandatory option.
   --name=NAME  Name of the video to upload. (default to video filename)
-  --debug  Trigger some debug information like options used (default: no)
   -d, --description=STRING  Description of the video. (default: default description)
   -t, --tags=STRING  Tags for the video. comma separated.
                      WARN: tags with punctuation (!, ', ", ?, ...)
@@ -49,6 +48,34 @@ Options:
   -h --help  Show this help.
   --version  Show version.
 
+Logging options
+  -q --quiet        Suppress any log except Critical (alias for --log=critical).
+  --log=STRING      Log level, between debug, info, warning, error, critical. Ignored if --quiet is set (default to info)
+  -u --url-only     Display generated URL after upload directly on stdout, implies --quiet
+  --batch           Display generated URL after upload with platform information for easier parsing. Implies --quiet
+                    Be careful --batch and --url-only are mutually exclusives.
+  --debug           (Deprecated) Alias for --log=debug. Ignored if --log is set
+
+Strict options:
+  Strict options allow you to force some option to be present when uploading a video. It's useful to be sure you do not
+  forget something when uploading a video, for example if you use multiples NFO. You may force the presence of description,
+  tags, thumbnail, ...
+  All strict option are optionals and are provided only to avoid errors when uploading :-)
+  All strict options can be specified in NFO directly, the only strict option mandatory on cli is --withNFO
+  All strict options are off by default
+
+  --withNFO         Prevent the upload without a NFO, either specified via cli or found in the directory
+  --withThumbnail       Prevent the upload without a thumbnail
+  --withName        Prevent the upload if no name are found
+  --withDescription     Prevent the upload without description
+  --withTags        Prevent the upload without tags
+  --withPlaylist    Prevent the upload if no playlist
+  --withPublishAt    Prevent the upload if no schedule
+  --withPlatform    Prevent the upload if at least one platform is not specified
+  --withCategory    Prevent the upload if no category
+  --withLanguage    Prevent upload if no language
+  --withChannel     Prevent upload if no channel
+
 Categories:
   Category is the type of video you upload. Default is films.
   Here are available categories from Peertube and Youtube:
@@ -69,9 +96,16 @@ import sys
 if sys.version_info[0] < 3:
     raise Exception("Python 3 or a more recent version is required.")
 
+import os
 import datetime
 import logging
-logging.basicConfig(format='%(asctime)s %(message)s', level=logging.INFO)
+logger = logging.getLogger('Prismedia')
+logger.setLevel(logging.INFO)
+ch = logging.StreamHandler()
+ch.setLevel(logging.INFO)
+formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s: %(message)s')
+ch.setFormatter(formatter)
+logger.addHandler(ch)
 
 from docopt import docopt
 
@@ -81,9 +115,9 @@ from . import utils
 
 try:
     # noinspection PyUnresolvedReferences
-    from schema import Schema, And, Or, Optional, SchemaError
+    from schema import Schema, And, Or, Optional, SchemaError, Hook, Use
 except ImportError:
-    logging.error('This program requires that the `schema` data-validation library'
+    logger.critical('This program requires that the `schema` data-validation library'
                   ' is installed: \n'
                   'see https://github.com/halst/schema\n')
     exit(1)
@@ -91,12 +125,12 @@ try:
     # noinspection PyUnresolvedReferences
     import magic
 except ImportError:
-    logging.error('This program requires that the `python-magic` library'
+    logger.critical('This program requires that the `python-magic` library'
                   ' is installed, NOT the Python bindings to libmagic API \n'
                   'see https://github.com/ahupp/python-magic\n')
     exit(1)
 
-VERSION = "prismedia v0.9.1"
+VERSION = "prismedia v0.10.0"
 
 VALID_PRIVACY_STATUSES = ('public', 'private', 'unlisted')
 VALID_CATEGORIES = (
@@ -170,17 +204,104 @@ def validatePublish(publish):
 
 def validateThumbnail(thumbnail):
     supported_types = ['image/jpg', 'image/jpeg']
-    if magic.from_file(thumbnail, mime=True) in supported_types:
+    if os.path.exists(thumbnail) and \
+            magic.from_file(thumbnail, mime=True) in supported_types:
         return thumbnail
     else:
         return False
 
 
+def validateLogLevel(loglevel):
+    numeric_level = getattr(logging, loglevel, None)
+    if not isinstance(numeric_level, int):
+        return False
+    return True
+
+def _optionnalOrStrict(key, scope, error):
+    option = key.replace('-', '')
+    option = option[0].upper() + option[1:]
+    if scope["--with" + option] is True and scope[key] is None:
+        logger.critical("Prismedia: you have required the strict presence of " + key + " but none is found")
+        exit(1)
+    return True
+
+
+def configureLogs(options):
+    if options.get('--batch') and options.get('--url-only'):
+        logger.critical("Prismedia: Please use either --batch OR --url-only, not both.")
+        exit(1)
+    # batch and url-only implies quiet
+    if options.get('--batch') or options.get('--url-only'):
+        options['--quiet'] = True
+
+    if options.get('--quiet'):
+        # We need to set both log level in the same time
+        logger.setLevel(50)
+        ch.setLevel(50)
+    elif options.get('--log'):
+        numeric_level = getattr(logging, options["--log"], None)
+        # We need to set both log level in the same time
+        logger.setLevel(numeric_level)
+        ch.setLevel(numeric_level)
+    elif options.get('--debug'):
+        logger.warning("DEPRECATION: --debug is deprecated, please use --log=debug instead")
+        logger.setLevel(10)
+        ch.setLevel(10)
+
+
+def configureStdoutLogs():
+    logger_stdout = logging.getLogger('stdoutlogs')
+    logger_stdout.setLevel(logging.INFO)
+    ch_stdout = logging.StreamHandler(stream=sys.stdout)
+    ch_stdout.setLevel(logging.INFO)
+    # Default stdout logs is url only
+    formatter_stdout = logging.Formatter('%(message)s')
+    ch_stdout.setFormatter(formatter_stdout)
+    logger_stdout.addHandler(ch_stdout)
+
 def main():
     options = docopt(__doc__, version=VERSION)
 
+    earlyoptionSchema = Schema({
+        Optional('--log'): Or(None, And(
+                                str,
+                                Use(str.upper),
+                                validateLogLevel,
+                                error="Log level not recognized")
+                              ),
+        Optional('--quiet', default=False): bool,
+        Optional('--debug'): bool,
+        Optional('--url-only', default=False): bool,
+        Optional('--batch', default=False): bool,
+        Optional('--withNFO', default=False): bool,
+        Optional('--withThumbnail', default=False): bool,
+        Optional('--withName', default=False): bool,
+        Optional('--withDescription', default=False): bool,
+        Optional('--withTags', default=False): bool,
+        Optional('--withPlaylist', default=False): bool,
+        Optional('--withPublishAt', default=False): bool,
+        Optional('--withPlatform', default=False): bool,
+        Optional('--withCategory', default=False): bool,
+        Optional('--withLanguage', default=False): bool,
+        Optional('--withChannel', default=False): bool,
+        # This allow to return all other options for further use: https://github.com/keleshev/schema#extra-keys
+        object: object
+    })
+
     schema = Schema({
-        '--file': And(str, validateVideo, error='file is not supported, please use mp4'),
+        '--file': And(str, os.path.exists, validateVideo, error='file is not supported, please use mp4'),
+        # Strict option checks - at the moment Schema needs to check Hook and Optional separately #
+        Hook('--name', handler=_optionnalOrStrict): object,
+        Hook('--description', handler=_optionnalOrStrict): object,
+        Hook('--tags', handler=_optionnalOrStrict): object,
+        Hook('--category', handler=_optionnalOrStrict): object,
+        Hook('--language', handler=_optionnalOrStrict): object,
+        Hook('--platform', handler=_optionnalOrStrict): object,
+        Hook('--publishAt', handler=_optionnalOrStrict): object,
+        Hook('--thumbnail', handler=_optionnalOrStrict): object,
+        Hook('--channel', handler=_optionnalOrStrict): object,
+        Hook('--playlist', handler=_optionnalOrStrict): object,
+        # Validate checks #
         Optional('--name'): Or(None, And(
                                 str,
                                 lambda x: not x.isdigit(),
@@ -228,7 +349,6 @@ def main():
                                     validatePublish,
                                     error="DATE should be the form YYYY-MM-DDThh:mm:ss and has to be in the future")
                                     ),
-        Optional('--debug'): bool,
         Optional('--cca'): bool,
         Optional('--disable-comments'): bool,
         Optional('--nsfw'): bool,
@@ -240,10 +360,29 @@ def main():
         Optional('--playlist'): Or(None, str),
         Optional('--playlistCreate'): bool,
         '--help': bool,
-        '--version': bool
+        '--version': bool,
+        # This allow to return all other options for further use: https://github.com/keleshev/schema#extra-keys
+        object: object
     })
+    # We need to validate early options first as withNFO and logs options should be prioritized
+    try:
+        options = earlyoptionSchema.validate(options)
+        configureLogs(options)
+    except SchemaError as e:
+        logger.critical(e)
+        exit(1)
+
+    if options.get('--url-only') or options.get('--batch'):
+        configureStdoutLogs()
 
     options = utils.parseNFO(options)
+
+    # Once NFO are loaded, we need to revalidate strict options in case some were in NFO
+    try:
+        options = earlyoptionSchema.validate(options)
+    except SchemaError as e:
+        logger.critical(e)
+        exit(1)
 
     if not options.get('--thumbnail'):
         options = utils.searchThumbnail(options)
@@ -251,11 +390,11 @@ def main():
     try:
         options = schema.validate(options)
     except SchemaError as e:
-        exit(e)
+        logger.critical(e)
+        exit(1)
 
-    if options.get('--debug'):
-        print(sys.version)
-        print(options)
+    logger.debug("Python " + sys.version)
+    logger.debug(options)
 
     if options.get('--platform') is None or "peertube" in options.get('--platform'):
         pt_upload.run(options)
@@ -264,6 +403,5 @@ def main():
 
 
 if __name__ == '__main__':
-    import warnings
-    warnings.warn("use 'python -m prismedia', not 'python -m prismedia.upload'", DeprecationWarning)
+    logger.warning("DEPRECATION: use 'python -m prismedia', not 'python -m prismedia.upload'")
     main()