+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## v0.10.0
+
+## Features
+ - Add the possibility to specify strict checks option to never forgot parameters when uploading (see #36)
+ - Improve logging system, add options for batch upload and print url-only in the stdout (see #29)
+ - --debug option is now deprecated in favor of --log=debug 
+ 
+## Fixes
+ - Workaround against the Youtube API breakdown while adding video in playlist. See #47 for details. Should be removed once Google fix their bugs.
+
+
 ## v0.9.1
 
 ### Features